chatbot/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "salesforce-bot-chatbot"
version = "0.1.0"
description = "Account Q&A Bot with Semantic Kernel and Azure services"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "gunicorn==21.2.0",
    "semantic-kernel==0.9.1b1",
    "openai==1.3.7",
    "azure-identity==1.15.0",
    "azure-cosmos==4.5.1",
    "azure-search-documents==11.4.0",
    "azure-keyvault-secrets==4.7.0",
    "azure-monitor-opentelemetry==1.2.0",
    "azure-storage-blob==12.19.0",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "numpy==1.25.2",
    "scipy==1.11.4",
    "gremlinpython==3.7.0",
    "pyodbc==5.0.1",
    "httpx==0.25.2",
    "python-multipart==0.0.6",
    "python-jose[cryptography]==3.3.0",
    "structlog==23.2.0",
    "tenacity==8.2.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0",
    "isort>=5.12",
    "flake8>=6.0",
    "mypy>=1.5",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
